# api/openapi.yaml
openapi: 3.1.0
info:
  title: dra-workflow API
  version: 0.1.0
  summary: Workflow-driven, AI-assisted IT service disaster recovery planning and execution.
  description: |
    REST API of the dra-workflow backend (Rust). It implements the data model in
    `docs/domain/dr-plan-model.md` and the workflows in `docs/workflows/`.

    **Hierarchy:** Tenant → IT service → microservice → DR items.

    **Methodology:** NIST SP 800-34 Rev. 1 and BSI-Standard 200-4 (see `docs/domain/standards-mapping.md`).

    ## Conventions
    - **Tenant from the token:** the tenant is **not** part of the path. The backend reads it from the
      `tenant_id` claim of the JWT access token and scopes every query to it. A token without a valid
      tenant claim, or for a tenant the user is not a member of, gets `403`. Ids belonging to another
      tenant return `404`. Only `GET/POST /tenants` (membership list, tenant creation) are tenant-independent.
    - **Collections** are nested under their parent (for example `/services/{serviceId}/microservices`).
      **Single items** are addressed flat by id (for example `/microservices/{microserviceId}`).
    - IDs are UUIDs. Durations are **integer minutes**. Timestamps are RFC 3339 UTC.
    - Updates use `PATCH` with `application/merge-patch+json` (RFC 7396).
    - Optimistic concurrency: every resource has a `version`. Reads return an `ETag`, and `PATCH`/`DELETE`
      accept `If-Match`. A stale version returns `412`.
...
      with `issues`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /services/{serviceId}/plan-versions:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    get:
      tags: [Plan Versions]
      operationId: listPlanVersions
      summary: List plan versions (newest first)
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PlanStatus' }
      responses:
        '200':
          description: Plan versions (without snapshot)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PlanVersionSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Plan Versions]
      operationId: submitPlanVersion
      summary: Snapshot the current service subtree and submit it for review (workflow step 12)
      description: Fails with `422` if any workflow gate from steps 1–11 has blocking issues.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanSubmission' }
      responses:
        '201':
          description: Plan version in status `in_review`
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersion' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /services/{serviceId}/handbook:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    get:
      tags: [Plan Versions]
      operationId: getDraftHandbook
      summary: Live preview of the emergency handbook from the current working data (draft, not approved)
      parameters:
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Markdown handbook marked as DRAFT
          content:
            text/markdown:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /plan-versions/{planVersionId}:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    get:
      tags: [Plan Versions]
      operationId: getPlanVersion
      summary: Get a plan version including its immutable snapshot
      responses:
        '200':
          description: Plan version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersion' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /plan-versions/{planVersionId}/approve:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    post:
      tags: [Plan Versions]
      operationId: approvePlanVersion
      summary: Approve (reviewer only). Retires the previously approved version.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanApproval' }
      responses:
        '200':
          description: Approved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /plan-versions/{planVersionId}/reject:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    post:
      tags: [Plan Versions]
      operationId: rejectPlanVersion
      summary: Reject a version in review and return it to draft
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlanRejection' }
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersion' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /plan-versions/{planVersionId}/export:
    parameters:
      - $ref: '#/components/parameters/PlanVersionId'
    get:
      tags: [Plan Versions]
      operationId: exportPlanVersion
      summary: Export as a standalone document for offline or crisis use (emergency handbook)
      description: |
        `markdown` and `json` are implemented; `pdf` returns `501` for now. The same export is available offline
        through the CLI: `dra-server export --tenant <slug> --plan-version <id>`.
      parameters:
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [markdown, pdf, json]
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Exported plan
          content:
            text/markdown:
              schema: { type: string }
            application/pdf:
              schema: { type: string, contentEncoding: binary }
            application/json:
              schema: { $ref: '#/components/schemas/PlanVersion' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── DR Tests ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token (JWT). Required claims: `sub` (user) and `tenant_id` (UUID of the active tenant).
        The tenant role may come from the token or from the tenant membership. The identity provider is
        still open (docs/overview/open-questions.md).
  headers:
    Location:
      description: URL of the created resource
      schema: { type: string }
...
  parameters:
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PlanVersionId:
      name: planVersionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AcceptLanguage:
      name: Accept-Language
      in: header
      description: Language for labels and exports (`en`, `de`)
      schema: { type: string, default: en }
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or it belongs to a different tenant than the token's)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The caller lacks the required tenant role or DR role, or the token has no valid `tenant_id` claim
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Conflicts with the current state (for example the resource is still referenced, or the transition is invalid)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationFailed:
      description: Input validation, integrity rule or workflow gate failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            issues:
              type: array
              items: { $ref: '#/components/schemas/Issue' }
    Issue:
      type: object
      description: A validation error, integrity-rule violation or gate failure
      required: [ruleId, severity, message]
      properties:
        ruleId:
          type: string
          description: Stable rule id, for example `RTO_EXCEEDS_SERVICE_RTO`, `DEPENDENCY_RTO_CONFLICT`, `STEP_WITHOUT_OWNER`, `STRATEGY_GAP`, `ROLE_WITHOUT_DEPUTY`
          examples: [DEPENDENCY_RTO_CONFLICT]
        severity: { type: string, enum: [blocking, warning, info] }
        message: { type: string }
        field: { type: string, description: 'JSON pointer to the offending input field, if any' }
        entityType: { type: string }
        entityId: { type: string, format: uuid }
        workflowStep: { $ref: '#/components/schemas/WorkflowStepKey' }
        standardRef:
          type: string
          description: Related standard reference
          examples: ['NIST SP 800-34 §3.2', 'BSI 200-4 Soll-Ist-Vergleich']
...
    ContentMeta:
      type: object
      properties:
        origin:
          type: string
          enum: [user, ai_accepted, import]
          readOnly: true
          description: Provenance of the content
        aiSuggestionId:
          type: [string, 'null']
          format: uuid
          readOnly: true
...
    Priority:
      type: string
      enum: [low, medium, high, critical]
...
    PlanStatus:
      type: string
      enum: [draft, in_review, approved, retired]
...
    Person:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/PersonFields'
        - required: [name]
...
    Role:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/RoleFields'
        - type: object
          required: [name]
          properties:
            isDefault: { type: boolean, readOnly: true }

    # ─────────── Catalog ───────────
...
    ItService:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/ItServiceFields'
        - type: object
          required: [name]
          properties:
            summary: { $ref: '#/components/schemas/ServiceSummary' }
...
    BusinessImpactAnalysis:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/BusinessImpactAnalysisInput'
        - type: object
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }
            derivedMtpdMinutes:
              type: [integer, 'null']
              readOnly: true
              description: MTPD derived from impact ratings and the tenant tolerance level (for plausibility checks)

    # ─────────── Scenario ───────────
...
    Scenario:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/ScenarioFields'
        - type: object
          required: [title, status]
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }
            status: { $ref: '#/components/schemas/ScenarioStatus' }
            mergedIntoId: { type: [string, 'null'], format: uuid, readOnly: true }
            likelihood: { $ref: '#/components/schemas/Rating' }
            impact: { $ref: '#/components/schemas/Rating' }
            riskScore: { type: integer, readOnly: true, description: likelihood × impact }
            priority: { $ref: '#/components/schemas/Priority' }
            drRequired: { $ref: '#/components/schemas/DrRequired' }
            decisionRationale: { type: string }
            decidedBy: { type: [string, 'null'], readOnly: true }
            decidedAt: { type: [string, 'null'], format: date-time, readOnly: true }

    # ─────────── Microservice ───────────
...
    Microservice:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/MicroserviceFields'
        - type: object
          required: [name]
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }
            isDefault:
              type: boolean
              readOnly: true
              description: The default component that stands for the whole service. It cannot be deleted.

    # ─────────── Dependency ───────────
...
    Dependency:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/DependencyFields'
        - type: object
          properties:
            microserviceId: { type: string, format: uuid, readOnly: true }
            rtoConflict:
              type: boolean
              readOnly: true
              description: True if this is a critical dependency whose RTO is greater than the dependent's RTO
...
    RecoveryObjective:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/RecoveryObjectiveFields'
        - type: object
          required: [rtoMinutes, rpoMinutes]
          properties:
            microserviceId: { type: string, format: uuid, readOnly: true }

    # ─────────── Recovery Strategy (Mitigation) ───────────
...
    RecoveryStrategy:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/RecoveryStrategyFields'
        - type: object
          required: [scenarioIds, type]
          properties:
            microserviceId: { type: string, format: uuid, readOnly: true }
            isSelected: { type: boolean, readOnly: true }
            gapJustification: { type: string, readOnly: true, description: Justification given when a gap was accepted }
            gapCheck: { $ref: '#/components/schemas/GapCheck' }
...
    DataProtection:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/DataProtectionFields'
        - type: object
          properties:
            microserviceId: { type: string, format: uuid, readOnly: true }
            supportsRpo:
              type: boolean
              readOnly: true
              description: True if the frequency is ≤ the microservice's RPO

    # ─────────── Runbook ───────────
...
    RunbookDetail:
      allOf:
        - $ref: '#/components/schemas/Runbook'
        - type: object
          properties:
            steps:
              type: array
              items: { $ref: '#/components/schemas/RunbookStep' }
...
    RoleAssignment:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/RoleAssignmentFields'
        - type: object
          required: [roleId, personId]
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }
...
    CommunicationRule:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/CommunicationRuleFields'
        - type: object
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }

    # ─────────── Workflow ───────────
...
    PlanVersionSummary:
      type: object
      required: [id, serviceId, version, status]
      properties:
        id: { type: string, format: uuid }
        serviceId: { type: string, format: uuid }
        version: { type: integer }
        status: { $ref: '#/components/schemas/PlanStatus' }
        submittedBy: { type: string }
        submittedAt: { type: string, format: date-time }
        approvedBy: { type: [string, 'null'] }
        approvedAt: { type: [string, 'null'], format: date-time }
        nextReviewDue: { type: [string, 'null'], format: date }
        comment: { type: string, description: Submission comment }
        reviewComment: { type: string, description: Comment of the approval or rejection }
    PlanVersion:
      allOf:
        - $ref: '#/components/schemas/PlanVersionSummary'
        - type: object
          properties:
            snapshot: { $ref: '#/components/schemas/PlanSnapshot' }
    PlanSnapshot:
      type: object
      description: Immutable copy of the full IT service subtree at submission time
      properties:
        formatVersion: { type: integer, description: Snapshot format version (currently 1) }
        service: { $ref: '#/components/schemas/ItService' }
        bia: { $ref: '#/components/schemas/BusinessImpactAnalysis' }
        scenarios:
          type: array
          items: { $ref: '#/components/schemas/Scenario' }
        microservices:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Microservice'
              - type: object
                properties:
                  dependencies:
                    type: array
                    items: { $ref: '#/components/schemas/Dependency' }
                  recoveryObjectives:
                    type: array
                    items: { $ref: '#/components/schemas/RecoveryObjective' }
                  recoveryStrategies:
                    type: array
                    items: { $ref: '#/components/schemas/RecoveryStrategy' }
                  dataProtection:
                    type: array
                    items: { $ref: '#/components/schemas/DataProtection' }
                  runbooks:
                    type: array
                    items: { $ref: '#/components/schemas/RunbookDetail' }
        roleAssignments:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/RoleAssignment'
              - type: object
                properties:
                  roleName: { type: string }
                  person: { $ref: '#/components/schemas/Person' }
        communicationRules:
          type: array
          items: { $ref: '#/components/schemas/CommunicationRule' }
        roles:
          type: array
          description: Role catalog of the tenant (to resolve role ids of runbook steps)
          items: { $ref: '#/components/schemas/Role' }
        persons:
          type: array
          description: Persons referenced by the plan (owners, role assignments)
          items: { $ref: '#/components/schemas/Person' }
    PlanSubmission:
      type: object
      properties:
        comment: { type: string }
    PlanApproval:
      type: object
      properties:
        comment: { type: string }
        nextReviewDue: { type: string, format: date, description: Defaults to approval date plus the tenant review interval }
    PlanRejection:
      type: object
      required: [comment]
      properties:
        comment: { type: string, minLength: 1 }

    # ─────────── Readiness ───────────
...